Project structure

+++ dat110-project1/src/no/hvl/dat110/messaging/MessageUtils.java
@@ -15,7 +15,7 @@ public class MessageUtils {
 
         byte[] segment = new byte[SEGMENTSIZE];
         byte[] data = message.getData();
-        segment[0] = data.length;
+        segment[0] = (byte) data.length;
 
         for (int i = 0; i < data.length; i++) {
             segment[i + 1] = data[i];
